Fix results length to match `max_papers`
We were assuming the results from the `collect_by_terms`
were unique, which is not true if the collector finds
repeated papers.

Instead of considering the length of the response data,
we are now using the length of unique papers actually
added to the resulting papers. This guarantees the
number of results will always match the `max_papers`
parameter.