Add support for including specific results in search API response
This commit adds support for including specific types of results in the search API response. A new `ResultsIncludes` enum is defined to specify which types of results should be included. The `Search.from_cls()` and `Results.from_cls()` methods are updated to accept an optional list of `ResultsIncludes` values, and only include the specified types of results in the output. The `/search/{search_id}` endpoint is updated to accept an optional query parameter specifying which types of results should be included in the response.