feat: Add stop functionality to collectors, expanders and Sciencer
This commit adds a `stop` method to the `Collector`, `Expander` and `Sciencer` classes. The method sets a `_stop` flag that is checked during execution of their respective methods. If set to True, the execution will be stopped immediately. Additionally, when the search is cancelled, the `stop` method of the Sciencer instance is called to stop all ongoing operations.